Add zustand persisted store
